refactor: extract spending comparison tooltip logic
